Fix compile error when using VS2012 toolchain.
BUG=0

Review URL: https://chromiumcodereview.appspot.com/22861009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@218349 0039d316-1c4b-4281-b951-d872f2087c98

+++ win8/metro_driver/file_picker_ash.cc
@@ -153,7 +153,7 @@ HRESULT OpenFilePickerSession::SinglePickerDone(SingleFileAsyncOp* async,
       LOG(ERROR) << "NULL IStorageItem";
     }
   } else {
-    LOG(ERROR) << "Unexpected async status " << status;
+    LOG(ERROR) << "Unexpected async status " << static_cast<int>(status);
   }
   app_view_->OnOpenFileCompleted(this, success_);
   return S_OK;
@@ -204,7 +204,7 @@ HRESULT OpenFilePickerSession::MultiPickerDone(MultiFileAsyncOp* async,
       LOG(ERROR) << "NULL StorageFileVectorCollection";
     }
   } else {
-    LOG(ERROR) << "Unexpected async status " << status;
+    LOG(ERROR) << "Unexpected async status " << static_cast<int>(status);
   }
   app_view_->OnOpenFileCompleted(this, success_);
   return S_OK;
@@ -544,7 +544,7 @@ HRESULT SaveFilePickerSession::FilePickerDone(SaveFileAsyncOp* async,
       LOG(ERROR) << "NULL IStorageItem";
     }
   } else {
-    LOG(ERROR) << "Unexpected async status " << status;
+    LOG(ERROR) << "Unexpected async status " << static_cast<int>(status);
   }
   app_view_->OnSaveFileCompleted(this, success_);
   return S_OK;
@@ -613,7 +613,7 @@ HRESULT FolderPickerSession::FolderPickerDone(FolderPickerAsyncOp* async,
       LOG(ERROR) << "NULL IStorageItem";
     }
   } else {
-    LOG(ERROR) << "Unexpected async status " << status;
+    LOG(ERROR) << "Unexpected async status " << static_cast<int>(status);
   }
   app_view_->OnFolderPickerCompleted(this, success_);
   return S_OK;
